+++ gems/dommy/test/test_js_wire_sync.rb
@@ -165,7 +165,7 @@ class TestProtoMethodStubsInvalidateCaches < Minitest::Test
 
   def test_the_stub_routes_a_mutating_call_through_the_epoch_bump
     assert_includes(RUNTIME_JS, "function callMutating(handle, name, wire)")
-    assert_match(/readOnly \? rehydrate\(__rb_host_call\(this\[HKEY\], name, wire\)\) : callMutating\(/, RUNTIME_JS)
+    assert_match(/readOnly \? hostCallResult\(name, __rb_host_call\(this\[HKEY\], name, wire\)\) : callMutating\(/, RUNTIME_JS)
   end
 
   def test_the_read_only_set_is_what_decides